Improved document & object fields by removing constructors
