--- ArticleTypes.ts
@@ -1,3 +1,0 @@
-export enum _ARTICLE_TYPES {
-	ARTICLE = 'article',
-}
