+++ DocumentTypes.ts
@@ -1,0 +1,15 @@
+export enum _ARTICLE_TYPES {
+	ARTICLE = 'article',
+}
+
+export enum _BLOCK_TYPES {
+	ARCHIVE = 'archive',
+	CONTACT = 'contact',
+	FEATURED_ARTICLES = 'featuredArticles',
+	FEATURED_TAXONOMIES = 'featuredTaxonomies',
+	GALLERY = 'gallery',
+	IMAGE = 'image',
+	INFO = 'info',
+	NEWSLETTER = 'newsletter',
+	TEXT = 'text',
+}
