+++ studio/sanity.config.ts
@@ -4,7 +4,7 @@
 import { structureTool } from 'sanity/structure'
 
 // Go to https://www.sanity.io/docs/api-versioning to learn how API versioning works
-import { apiVersion, configStudioBasePath, configStudioName, configStudioTitle, dataset, projectId } from './src/lib/api'
+import { apiVersion, configStudioBasePath, configStudioName, configStudioTitle, dataset, projectId } from '@lib.studio/api'
 import { structure, schemaOptions, documentOptions } from './src/structure'
 import { PluginOptions, StudioToolMenu, defineConfig } from 'sanity'
 import { media } from 'sanity-plugin-media'
