+++ studio/src/schemas/documents/pages/archive.tsx
@@ -1,6 +1,16 @@
-import { page } from './utils/page';
+import { _PAGE_FIELDS } from "@schemas.studio/documents/pages/utils/fields"
+import { _PAGE_PREVIEW } from "@schemas.studio/documents/pages/utils/preview"
+import { defineField, defineType } from "sanity"
 
-const fields = [
+const _FIELDS: any = [
 ]
 
-export const archive = page({ name: 'archive' })
+export const archive = defineType({ 
+	name: 'archive',
+	type: 'document',
+	fields: [
+		..._PAGE_FIELDS,
+		..._FIELDS
+	],
+	preview: _PAGE_PREVIEW
+ })
