--- studio/src/schemas/documents/pages/blocks/Contact.tsx
@@ -1,14 +1,0 @@
-import { defineField } from 'sanity';
-import constructors from '../utils';
-
-import { InboxArrowDownIcon } from '@heroicons/react/24/solid'
-
-const fields: any = [
-	defineField({
-		title: 'Button Text',
-		name: 'buttonText',
-		type: 'string',
-	}),
-]
-
-export const Contact = constructors.block({ name: 'Contact', fields, Icon: InboxArrowDownIcon })
