--- studio/src/schemas/documents/pages/utils/index.ts
@@ -1,4 +1,0 @@
-import { block } from "./block"
-
-export default { block }
-
