+++ studio/src/schemas/objects/blocks/utils/fields.tsx
@@ -1,0 +1,2 @@
+export const _BLOCK_FIELDS = [
+]
